+++ ClassicLootManager/Modules/Common/ConfigManager/ConfigManager.lua
@@ -127,7 +127,7 @@ end
 ---@param append boolean?
 function ConfigManager:AddGroup(groupName, append)
     if CONSTANTS.CONFIGS.GROUPS[groupName] then
-        error("Config group %s already exists.", groupName)
+        error(string.format("Config group %s already exists.", groupName))
         return
     end
     -- Never allow changing first group